+++ src/main/java/com/ebicep/warlords/pve/upgrades/paladin/avenger/AvengerStrikeBranch.java
@@ -49,7 +49,7 @@ public class AvengerStrikeBranch extends AbstractUpgradeBranch<AvengersStrike> {
         ability.setMaxDamageHeal(maxDamage * 1.6f);
     }
 
-    int energyCost = ability.getEnergyCost();
+    float energyCost = ability.getEnergyCost();
 
     @Override
     public void b1() {
